+++ src/query/frontend/stripped_lexer_constants.hpp
@@ -79,14 +79,14 @@ class Trie {
 const int kBitsetSize = 65536;
 
 const trie::Trie kKeywords = {
-    "union",  "all",     "optional",  "match",  "unwind",     "as",
-    "merge",  "on",      "create",    "set",    "detach",     "delete",
-    "remove", "with",    "distinct",  "return", "order",      "by",
-    "skip",   "limit",   "ascending", "asc",    "descending", "desc",
-    "where",  "or",      "xor",       "and",    "not",        "in",
-    "starts", "ends",    "contains",  "is",     "null",       "count",
-    "filter", "extract", "any",       "none",   "single",     "true",
-    "false"};
+    "union",   "all",   "optional",  "match",  "unwind",     "as",
+    "merge",   "on",    "create",    "set",    "detach",     "delete",
+    "remove",  "with",  "distinct",  "return", "order",      "by",
+    "skip",    "limit", "ascending", "asc",    "descending", "desc",
+    "where",   "or",    "xor",       "and",    "not",        "in",
+    "starts",  "ends",  "contains",  "is",     "null",       "case",
+    "when",    "then",  "else",      "end",    "count",      "filter",
+    "extract", "any",   "none",      "single", "true",       "false"};
 
 // Unicode codepoints that are allowed at the start of the unescaped name.
 const std::bitset<kBitsetSize> kUnescapedNameAllowedStarts(std::string(